requirements.txt
fastkml[lxml]>=1.4.0
lxml>=5.0
geopandas>=0.14
pandapower>=3.4.0
pandas>=2.0
numpy>=1.24
scipy>=1.10
requests>=2.31
pyyaml>=6.0
openpyxl>=3.1
pulp>=3.3.0
highspy>=1.7.0  # HiGHS solver backend for better MIP performance
tqdm>=4.65
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
pytest>=7.4
pytest-cov>=4.1
